feat(adr-0005): split ucal.rs into shared_buffer + metal_compute
- src/shared_buffer.rs: SharedBuffer, canonical ComputeBackend{Cuda(usize),
  Rocm(usize),Metal,Vulkan,Cpu} with is_gpu/device_ordinal/name; unit tests
- src/metal_compute.rs: MetalError, MetalDtype, MetalBuffer, ThreadgroupSize,
  GridSize, MetalKernel, MetalKernelLibrary, MetalContext, CommandOp,
  CommandBufferEncoder, CommandBuffer, ExecutionStatus, ExecutionHandle,
  TieredExecutor — all pure-Rust simulated Metal infrastructure
- src/ucal.rs: compatibility re-export shim (all existing imports still work)
- src/lib.rs: registered shared_buffer, metal_compute alongside ucal shim

cargo check: 0 errors
