+++ src/ucal.rs
@@ -1,1629 +1,15 @@
-//! Metal Compute Backend for macOS (Apple Silicon).
+//! Compatibility shim — `ucal` re-exports from the split modules.
 //!
-//! Provides GPU-accelerated compute operations via Apple's Metal API.
-//! This is the macOS equivalent of the CUDA backend — all matrix ops,
-//! attention kernels, and quantised inference run on the Apple GPU.
+//! `ucal.rs` has been split into:
+//! - `shared_buffer.rs`  — platform-agnostic `SharedBuffer` + `ComputeBackend`
+//! - `metal_compute.rs`  — Metal kernels, context, command encoding
 //!
-//! Architecture:
-//!   - `MetalDevice`: Wraps an MTLDevice for resource allocation & command submission
-//!   - `MetalBuffer`: GPU buffer wrapper with typed accessors
-//!   - `MetalKernel`: Compiled compute pipeline (MSL shader)
-//!   - `MetalContext`: High-level context managing device + command queue
-//!
-//! Feature-gated: compile with `--features metal` to enable.
-
-use std::collections::HashMap;
-use std::fmt;
-
-// ---------------------------------------------------------------------------
-// Error handling
-// ---------------------------------------------------------------------------
-
-/// Metal-specific errors.
-#[derive(Debug)]
-pub enum MetalError {
-    /// No Metal-capable GPU found on this system.
-    DeviceNotFound,
-    /// Failed to create a Metal buffer.
-    BufferCreation { size: usize, msg: String },
-    /// Kernel compilation failure.
-    KernelCompilation { name: String, error: String },
-    /// Command buffer submission failure.
-    CommandSubmission(String),
-    /// Attempted to use an unsupported dtype.
-    UnsupportedDtype(String),
-    /// Buffer is too small for the requested operation.
-    BufferTooSmall { required: usize, actual: usize },
-    /// Invalid kernel dispatch dimensions.
-    InvalidDispatch { reason: String },
-}
-
-impl fmt::Display for MetalError {
-    fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
-        match self {
-            MetalError::DeviceNotFound => write!(f, "No Metal-capable GPU found"),
-            MetalError::BufferCreation { size, msg } => {
-                write!(f, "Failed to create Metal buffer of {} bytes: {}", size, msg)
-            }
-            MetalError::KernelCompilation { name, error } => {
-                write!(f, "Metal kernel '{}' compilation failed: {}", name, error)
-            }
-            MetalError::CommandSubmission(msg) => {
-                write!(f, "Metal command submission failed: {}", msg)
-            }
-            MetalError::UnsupportedDtype(d) => {
-                write!(f, "Unsupported dtype for Metal compute: {}", d)
-            }
-            MetalError::BufferTooSmall { required, actual } => {
-                write!(
-                    f,
-                    "Buffer too small: need {} bytes, have {} bytes",
-                    required, actual
-                )
-            }
-            MetalError::InvalidDispatch { reason } => {
-                write!(f, "Invalid dispatch dimensions: {}", reason)
-            }
-        }
-    }
-}
-
-impl std::error::Error for MetalError {}
-
-pub type MetalResult<T> = Result<T, MetalError>;
-
-// ---------------------------------------------------------------------------
-// Data types for Metal compute
-// ---------------------------------------------------------------------------
-
-/// Supported dtypes for Metal buffers.
-#[derive(Debug, Clone, Copy, PartialEq, Eq, Hash)]
-pub enum MetalDtype {
-    Float32,
-    Float16,
-    BFloat16,
-    Int8,
-    Int32,
-    Uint8,
-    Uint32,
-}
-
-impl MetalDtype {
-    /// Size of one element in bytes.
-    pub fn size_bytes(&self) -> usize {
-        match self {
-            MetalDtype::Float32 => 4,
-            MetalDtype::Float16 => 2,
-            MetalDtype::BFloat16 => 2,
-            MetalDtype::Int8 | MetalDtype::Uint8 => 1,
-            MetalDtype::Int32 | MetalDtype::Uint32 => 4,
-        }
-    }
-
-    /// Metal Shading Language type name.
-    pub fn msl_type(&self) -> &'static str {
-        match self {
-            MetalDtype::Float32 => "float",
-            MetalDtype::Float16 => "half",
-            MetalDtype::BFloat16 => "bfloat",
-            MetalDtype::Int8 => "char",
-            MetalDtype::Int32 => "int",
-            MetalDtype::Uint8 => "uchar",
-            MetalDtype::Uint32 => "uint",
-        }
-    }
-}
-
-impl fmt::Display for MetalDtype {
-    fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
-        write!(f, "{}", self.msl_type())
-    }
-}
-
-// ---------------------------------------------------------------------------
-// Metal Buffer
-// ---------------------------------------------------------------------------
-
-/// A Metal GPU buffer with size tracking and dtype metadata.
-pub struct MetalBuffer {
-    /// Unique buffer ID (for tracking).
-    pub id: u64,
-    /// Size in bytes.
-    pub size_bytes: usize,
-    /// Number of typed elements.
-    pub element_count: usize,
-    /// Element dtype.
-    pub dtype: MetalDtype,
-    /// CPU-side shadow for readback.
-    data: Vec<u8>,
-}
-
-impl MetalBuffer {
-    /// Create a new Metal buffer.
-    pub fn new(element_count: usize, dtype: MetalDtype, id: u64) -> MetalResult<Self> {
-        let size_bytes = element_count * dtype.size_bytes();
-        if size_bytes == 0 {
-            return Err(MetalError::BufferCreation {
-                size: 0,
-                msg: "Cannot create zero-size buffer".to_string(),
-            });
-        }
-
-        Ok(Self {
-            id,
-            size_bytes,
-            element_count,
-            dtype,
-            data: vec![0u8; size_bytes],
-        })
-    }
-
-    /// Create from existing f32 data.
-    pub fn from_f32(data: &[f32], id: u64) -> MetalResult<Self> {
-        let mut buf = Self::new(data.len(), MetalDtype::Float32, id)?;
-        let bytes = unsafe {
-            std::slice::from_raw_parts(data.as_ptr() as *const u8, data.len() * 4)
-        };
-        buf.data.copy_from_slice(bytes);
-        Ok(buf)
-    }
-
-    /// Read buffer contents as f32 slice.
-    pub fn as_f32(&self) -> MetalResult<&[f32]> {
-        if self.dtype != MetalDtype::Float32 {
-            return Err(MetalError::UnsupportedDtype(format!(
-                "Buffer is {:?}, not Float32",
-                self.dtype
-            )));
-        }
-        Ok(unsafe {
-            std::slice::from_raw_parts(self.data.as_ptr() as *const f32, self.element_count)
-        })
-    }
-
-    /// Write data to buffer.
-    pub fn write_f32(&mut self, data: &[f32]) -> MetalResult<()> {
-        if self.dtype != MetalDtype::Float32 {
-            return Err(MetalError::UnsupportedDtype(format!(
-                "Buffer is {:?}, not Float32",
-                self.dtype
-            )));
-        }
-
-        let needed = data.len() * 4;
-        if needed > self.size_bytes {
-            return Err(MetalError::BufferTooSmall {
-                required: needed,
-                actual: self.size_bytes,
-            });
-        }
-
-        let bytes = unsafe {
-            std::slice::from_raw_parts(data.as_ptr() as *const u8, needed)
-        };
-        self.data[..needed].copy_from_slice(bytes);
-        Ok(())
-    }
-
-    /// Raw byte access.
-    pub fn as_bytes(&self) -> &[u8] {
-        &self.data
-    }
-}
-
-impl fmt::Debug for MetalBuffer {
-    fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
-        f.debug_struct("MetalBuffer")
-            .field("id", &self.id)
-            .field("dtype", &self.dtype)
-            .field("elements", &self.element_count)
-            .field("bytes", &self.size_bytes)
-            .finish()
-    }
-}
-
-// ---------------------------------------------------------------------------
-// SharedBuffer (backward-compatible with pipeline.rs)
-// ---------------------------------------------------------------------------
-
-/// A shared buffer for cross-component use (e.g., pipeline slots).
-///
-/// Wraps a reference-counted byte slice so pipeline slots can hold
-/// weight data without copying.
-#[derive(Debug, Clone)]
-pub struct SharedBuffer {
-    /// Raw byte data (used for layer weight storage).
-    pub data: std::sync::Arc<Vec<u8>>,
-    /// Size in bytes.
-    pub size: usize,
-}
-
-impl SharedBuffer {
-    /// Create from raw bytes.
-    pub fn from_bytes(data: Vec<u8>) -> Self {
-        let size = data.len();
-        Self {
-            data: std::sync::Arc::new(data),
-            size,
-        }
-    }
-
-    /// Create an empty buffer of given size.
-    pub fn zeroed(size: usize) -> Self {
-        Self {
-            data: std::sync::Arc::new(vec![0u8; size]),
-            size,
-        }
-    }
-
-    /// Borrow the underlying bytes.
-    pub fn as_bytes(&self) -> &[u8] {
-        &self.data
-    }
-}
-
-// ---------------------------------------------------------------------------
-// Compute Kernel
-// ---------------------------------------------------------------------------
-
-/// Threadgroup dimensions for Metal dispatch.
-#[derive(Debug, Clone, Copy)]
-pub struct ThreadgroupSize {
-    pub width: u32,
-    pub height: u32,
-    pub depth: u32,
-}
-
-impl ThreadgroupSize {
-    pub fn new_1d(width: u32) -> Self {
-        Self { width, height: 1, depth: 1 }
-    }
-
-    pub fn new_2d(width: u32, height: u32) -> Self {
-        Self { width, height, depth: 1 }
-    }
-
-    pub fn total_threads(&self) -> u32 {
-        self.width * self.height * self.depth
-    }
-}
-
-/// Grid dimensions for Metal dispatch.
-#[derive(Debug, Clone, Copy)]
-pub struct GridSize {
-    pub width: u32,
-    pub height: u32,
-    pub depth: u32,
-}
-
-impl GridSize {
-    pub fn new_1d(width: u32) -> Self {
-        Self { width, height: 1, depth: 1 }
-    }
-
-    pub fn new_2d(width: u32, height: u32) -> Self {
-        Self { width, height, depth: 1 }
-    }
-}
-
-/// A compiled Metal compute kernel.
-pub struct MetalKernel {
-    /// Function name in the MSL source.
-    pub name: String,
-    /// MSL source code.
-    pub source: String,
-    /// Recommended threadgroup size.
-    pub threadgroup_size: ThreadgroupSize,
-    /// Whether the kernel is compiled.
-    pub is_compiled: bool,
-}
-
-impl MetalKernel {
-    /// Create a new kernel (not yet compiled).
-    pub fn new(name: &str, source: &str, threadgroup_size: ThreadgroupSize) -> Self {
-        Self {
-            name: name.to_string(),
-            source: source.to_string(),
-            threadgroup_size,
-            is_compiled: false,
-        }
-    }
-
-    /// Simulate compilation (in real impl, calls MTLDevice::newLibraryWithSource).
-    pub fn compile(&mut self) -> MetalResult<()> {
-        if self.source.is_empty() {
-            return Err(MetalError::KernelCompilation {
-                name: self.name.clone(),
-                error: "Empty source".to_string(),
-            });
-        }
-        self.is_compiled = true;
-        Ok(())
-    }
-
-    /// Compute grid size for a 1D dispatch of `n` elements.
-    pub fn grid_for_elements(&self, n: u32) -> GridSize {
-        let groups = (n + self.threadgroup_size.width - 1) / self.threadgroup_size.width;
-        GridSize::new_1d(groups * self.threadgroup_size.width)
-    }
-}
-
-impl fmt::Debug for MetalKernel {
-    fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
-        f.debug_struct("MetalKernel")
-            .field("name", &self.name)
-            .field("compiled", &self.is_compiled)
-            .field("tg_size", &self.threadgroup_size.total_threads())
-            .finish()
-    }
-}
-
-// ---------------------------------------------------------------------------
-// Metal Context (high-level API)
-// ---------------------------------------------------------------------------
-
-/// High-level Metal compute context managing device, buffers, and kernels.
-///
-/// This is the primary entry point for the Air.rs Metal backend.
-pub struct MetalContext {
-    /// Device name (e.g., "Apple M2 Max").
-    pub device_name: String,
-    /// Maximum buffer size (typically 256GB on Apple Silicon).
-    pub max_buffer_size: usize,
-    /// Unified memory size in bytes.
-    pub unified_memory_bytes: usize,
-    /// Registered kernels by name.
-    kernels: HashMap<String, MetalKernel>,
-    /// Buffer ID counter.
-    next_buffer_id: u64,
-    /// Total allocated bytes.
-    allocated_bytes: usize,
-}
-
-impl MetalContext {
-    /// Create a new Metal context (simulated).
-    ///
-    /// In production, this calls `MTLCreateSystemDefaultDevice()`.
-    pub fn new() -> MetalResult<Self> {
-        Self::with_device("Apple M-Series GPU (simulated)")
-    }
-
-    /// Create with a specific device name.
-    pub fn with_device(device_name: &str) -> MetalResult<Self> {
-        Ok(Self {
-            device_name: device_name.to_string(),
-            max_buffer_size: 256 * 1024 * 1024 * 1024, // 256GB for Apple Silicon
-            unified_memory_bytes: 0,
-            kernels: HashMap::new(),
-            next_buffer_id: 1,
-            allocated_bytes: 0,
-        })
-    }
-
-    /// Allocate a new buffer on the Metal device.
-    pub fn alloc_buffer(
-        &mut self,
-        element_count: usize,
-        dtype: MetalDtype,
-    ) -> MetalResult<MetalBuffer> {
-        let id = self.next_buffer_id;
-        self.next_buffer_id += 1;
-
-        let buf = MetalBuffer::new(element_count, dtype, id)?;
-        self.allocated_bytes += buf.size_bytes;
-        Ok(buf)
-    }
-
-    /// Create a buffer from f32 data.
-    pub fn alloc_f32(&mut self, data: &[f32]) -> MetalResult<MetalBuffer> {
-        let id = self.next_buffer_id;
-        self.next_buffer_id += 1;
-
-        let buf = MetalBuffer::from_f32(data, id)?;
-        self.allocated_bytes += buf.size_bytes;
-        Ok(buf)
-    }
-
-    /// Free a buffer (returns its size for accounting).
-    pub fn free_buffer(&mut self, buf: MetalBuffer) -> usize {
-        let freed = buf.size_bytes;
-        self.allocated_bytes = self.allocated_bytes.saturating_sub(freed);
-        freed
-    }
-
-    /// Register and compile a compute kernel.
-    pub fn register_kernel(
-        &mut self,
-        name: &str,
-        source: &str,
-        threadgroup_size: ThreadgroupSize,
-    ) -> MetalResult<()> {
-        let mut kernel = MetalKernel::new(name, source, threadgroup_size);
-        kernel.compile()?;
-        self.kernels.insert(name.to_string(), kernel);
-        Ok(())
-    }
-
-    /// Get a registered kernel by name.
-    pub fn get_kernel(&self, name: &str) -> Option<&MetalKernel> {
-        self.kernels.get(name)
-    }
-
-    /// Dispatch a compute kernel (simulated).
-    ///
-    /// In production, this encodes into an MTLCommandBuffer and commits.
-    pub fn dispatch(
-        &self,
-        kernel_name: &str,
-        grid: GridSize,
-        _buffers: &[&MetalBuffer],
-    ) -> MetalResult<()> {
-        let kernel = self.kernels.get(kernel_name).ok_or_else(|| {
-            MetalError::KernelCompilation {
-                name: kernel_name.to_string(),
-                error: "Kernel not registered".to_string(),
-            }
-        })?;
-
-        if !kernel.is_compiled {
-            return Err(MetalError::KernelCompilation {
-                name: kernel_name.to_string(),
-                error: "Kernel not compiled".to_string(),
-            });
-        }
-
-        if grid.width == 0 || grid.height == 0 || grid.depth == 0 {
-            return Err(MetalError::InvalidDispatch {
-                reason: "Grid dimension must be > 0".to_string(),
-            });
-        }
-
-        // In production: encode → commit → waitUntilCompleted
-        Ok(())
-    }
-
-    /// Execute a simple element-wise f32 operation (simulated).
-    ///
-    /// This demonstrates the end-to-end flow:
-    /// 1. Create buffers
-    /// 2. Register kernel
-    /// 3. Dispatch
-    /// 4. Read results
-    pub fn elementwise_add(
-        &mut self,
-        a: &[f32],
-        b: &[f32],
-    ) -> MetalResult<Vec<f32>> {
-        if a.len() != b.len() {
-            return Err(MetalError::InvalidDispatch {
-                reason: format!("Mismatched lengths: {} vs {}", a.len(), b.len()),
-            });
-        }
-
-        // Register the add kernel if not present
-        if !self.kernels.contains_key("elementwise_add") {
-            self.register_kernel(
-                "elementwise_add",
-                r#"
-                #include <metal_stdlib>
-                using namespace metal;
-                kernel void elementwise_add(
-                    device const float* a [[buffer(0)]],
-                    device const float* b [[buffer(1)]],
-                    device float* out     [[buffer(2)]],
-                    uint id               [[thread_position_in_grid]])
-                {
-                    out[id] = a[id] + b[id];
-                }
-                "#,
-                ThreadgroupSize::new_1d(256),
-            )?;
-        }
-
-        let buf_a = self.alloc_f32(a)?;
-        let buf_b = self.alloc_f32(b)?;
-        let mut buf_out = self.alloc_buffer(a.len(), MetalDtype::Float32)?;
-
-        let kernel = self.kernels.get("elementwise_add").unwrap();
-        let grid = kernel.grid_for_elements(a.len() as u32);
-        self.dispatch("elementwise_add", grid, &[&buf_a, &buf_b, &buf_out])?;
-
-        // Simulate: compute on CPU (in production, GPU does this)
-        let result: Vec<f32> = a.iter().zip(b.iter()).map(|(x, y)| x + y).collect();
-        buf_out.write_f32(&result)?;
-
-        let output = buf_out.as_f32()?.to_vec();
-
-        self.free_buffer(buf_a);
-        self.free_buffer(buf_b);
-        self.free_buffer(buf_out);
-
-        Ok(output)
-    }
-
-    /// Total bytes currently allocated.
-    pub fn allocated_bytes(&self) -> usize {
-        self.allocated_bytes
-    }
-
-    /// Number of registered kernels.
-    pub fn kernel_count(&self) -> usize {
-        self.kernels.len()
-    }
-}
-
-impl fmt::Debug for MetalContext {
-    fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
-        f.debug_struct("MetalContext")
-            .field("device", &self.device_name)
-            .field("allocated_bytes", &self.allocated_bytes)
-            .field("kernels", &self.kernels.len())
-            .finish()
-    }
-}
-
-// ---------------------------------------------------------------------------
-// Built-in MSL kernel sources for common LLM ops
-// ---------------------------------------------------------------------------
-
-/// Pre-defined MSL kernel sources for inference operations.
-pub struct MetalKernelLibrary;
-
-impl MetalKernelLibrary {
-    /// Softmax kernel source (1D, per-row).
-    pub fn softmax_source() -> &'static str {
-        r#"
-        #include <metal_stdlib>
-        using namespace metal;
-        kernel void softmax(
-            device const float* input  [[buffer(0)]],
-            device float* output       [[buffer(1)]],
-            constant uint& row_size    [[buffer(2)]],
-            uint id                    [[thread_position_in_grid]])
-        {
-            uint row = id;
-            uint offset = row * row_size;
-
-            // Find max for numerical stability
-            float max_val = input[offset];
-            for (uint i = 1; i < row_size; i++) {
-                max_val = max(max_val, input[offset + i]);
-            }
-
-            // Compute exp and sum
-            float sum = 0.0;
-            for (uint i = 0; i < row_size; i++) {
-                float e = exp(input[offset + i] - max_val);
-                output[offset + i] = e;
-                sum += e;
-            }
-
-            // Normalise
-            for (uint i = 0; i < row_size; i++) {
-                output[offset + i] /= sum;
-            }
-        }
-        "#
-    }
-
-    /// RMS Norm kernel source.
-    pub fn rmsnorm_source() -> &'static str {
-        r#"
-        #include <metal_stdlib>
-        using namespace metal;
-        kernel void rmsnorm(
-            device const float* input  [[buffer(0)]],
-            device const float* weight [[buffer(1)]],
-            device float* output       [[buffer(2)]],
-            constant uint& dim         [[buffer(3)]],
-            constant float& eps        [[buffer(4)]],
-            uint id                    [[thread_position_in_grid]])
-        {
-            uint offset = id * dim;
-
-            // Compute mean of squares
-            float sum_sq = 0.0;
-            for (uint i = 0; i < dim; i++) {
-                float v = input[offset + i];
-                sum_sq += v * v;
-            }
-            float rms = rsqrt(sum_sq / float(dim) + eps);
-
-            // Normalise and scale
-            for (uint i = 0; i < dim; i++) {
-                output[offset + i] = input[offset + i] * rms * weight[i];
-            }
-        }
-        "#
-    }
-
-    /// SILU (SwiGLU) activation kernel source.
-    pub fn silu_source() -> &'static str {
-        r#"
-        #include <metal_stdlib>
-        using namespace metal;
-        kernel void silu(
-            device const float* input [[buffer(0)]],
-            device float* output      [[buffer(1)]],
-            uint id                   [[thread_position_in_grid]])
-        {
-            float x = input[id];
-            output[id] = x / (1.0 + exp(-x));
-        }
-        "#
-    }
-
-    /// RoPE (Rotary Positional Embedding) kernel source.
-    pub fn rope_source() -> &'static str {
-        r#"
-        #include <metal_stdlib>
-        using namespace metal;
-        kernel void rope(
-            device float* x           [[buffer(0)]],
-            constant uint& dim        [[buffer(1)]],
-            constant uint& pos        [[buffer(2)]],
-            constant float& theta     [[buffer(3)]],
-            uint id                   [[thread_position_in_grid]])
-        {
-            uint half_dim = dim / 2;
-            if (id >= half_dim) return;
-
-            float freq = 1.0 / pow(theta, float(2 * id) / float(dim));
-            float angle = float(pos) * freq;
-            float cos_a = cos(angle);
-            float sin_a = sin(angle);
-
-            float x0 = x[id];
-            float x1 = x[id + half_dim];
-            x[id]            = x0 * cos_a - x1 * sin_a;
-            x[id + half_dim] = x0 * sin_a + x1 * cos_a;
-        }
-        "#
-    }
-}
-
-// ---------------------------------------------------------------------------
-// Command Buffer Encoder — record → seal → dispatch pattern
-// ---------------------------------------------------------------------------
-
-/// Compute operations that can be recorded into a command buffer.
-///
-/// Each variant captures the full parameter set needed to dispatch on any
-/// backend (Metal, CPU fallback, or remote).
-#[derive(Debug, Clone)]
-pub enum CommandOp {
-    /// Dense matrix multiply: C = A × B.
-    /// (rows_a, inner_dim, cols_b, dtype)
-    MatMul {
-        rows_a: usize,
-        inner: usize,
-        cols_b: usize,
-        dtype: MetalDtype,
-    },
-    /// Row-wise softmax normalisation.
-    /// (rows, row_size)
-    Softmax { rows: usize, row_size: usize },
-    /// Layer normalisation with learned affine: y = γ·(x - μ)/σ + β.
-    /// (batch, dim, eps)
-    LayerNorm {
-        batch: usize,
-        dim: usize,
-        eps: f32,
-    },
-    /// RMS normalisation: y = x / rms(x) · weight.
-    /// (batch, dim, eps)
-    RmsNorm {
-        batch: usize,
-        dim: usize,
-        eps: f32,
-    },
-    /// Rotary Positional Embedding.
-    /// (seq_len, head_dim, base_theta)
-    RoPE {
-        seq_len: usize,
-        head_dim: usize,
-        base_theta: f32,
-    },
-    /// Device-to-device memory copy.
-    /// (num_bytes)
-    Copy { num_bytes: usize },
-}
-
-impl CommandOp {
-    /// Human-readable op name for profiling / debug.
-    pub fn name(&self) -> &'static str {
-        match self {
-            CommandOp::MatMul { .. } => "matmul",
-            CommandOp::Softmax { .. } => "softmax",
-            CommandOp::LayerNorm { .. } => "layer_norm",
-            CommandOp::RmsNorm { .. } => "rms_norm",
-            CommandOp::RoPE { .. } => "rope",
-            CommandOp::Copy { .. } => "copy",
-        }
-    }
-
-    /// Estimated FLOPs for this op (order-of-magnitude, for scheduling).
-    pub fn estimated_flops(&self) -> u64 {
-        match self {
-            CommandOp::MatMul {
-                rows_a,
-                inner,
-                cols_b,
-                ..
-            } => 2 * (*rows_a as u64) * (*inner as u64) * (*cols_b as u64),
-            CommandOp::Softmax { rows, row_size } => 5 * (*rows as u64) * (*row_size as u64),
-            CommandOp::LayerNorm { batch, dim, .. } => 5 * (*batch as u64) * (*dim as u64),
-            CommandOp::RmsNorm { batch, dim, .. } => 4 * (*batch as u64) * (*dim as u64),
-            CommandOp::RoPE {
-                seq_len, head_dim, ..
-            } => 6 * (*seq_len as u64) * (*head_dim as u64),
-            CommandOp::Copy { num_bytes } => *num_bytes as u64,
-        }
-    }
-}
-
-/// Encoder for recording operations before submission.
-///
-/// Operations are appended via `record_*` methods. Call `seal()` to
-/// produce a frozen `CommandBuffer` ready for dispatch.
-///
-/// Encoder is single-use: once sealed it cannot accept more ops.
-pub struct CommandBufferEncoder {
-    ops: Vec<CommandOp>,
-    sealed: bool,
-    label: String,
-}
-
-impl CommandBufferEncoder {
-    /// Create a new encoder with a debug label.
-    pub fn new(label: &str) -> Self {
-        Self {
-            ops: Vec::new(),
-            sealed: false,
-            label: label.to_string(),
-        }
-    }
-
-    /// Record a dense matrix multiplication.
-    pub fn record_matmul(
-        &mut self,
-        rows_a: usize,
-        inner: usize,
-        cols_b: usize,
-        dtype: MetalDtype,
-    ) -> Result<(), MetalError> {
-        self.check_writable()?;
-        self.ops.push(CommandOp::MatMul {
-            rows_a,
-            inner,
-            cols_b,
-            dtype,
-        });
-        Ok(())
-    }
-
-    /// Record a softmax normalisation.
-    pub fn record_softmax(&mut self, rows: usize, row_size: usize) -> Result<(), MetalError> {
-        self.check_writable()?;
-        self.ops.push(CommandOp::Softmax { rows, row_size });
-        Ok(())
-    }
-
-    /// Record a layer normalisation.
-    pub fn record_layer_norm(
-        &mut self,
-        batch: usize,
-        dim: usize,
-        eps: f32,
-    ) -> Result<(), MetalError> {
-        self.check_writable()?;
-        self.ops.push(CommandOp::LayerNorm { batch, dim, eps });
-        Ok(())
-    }
-
-    /// Record an RMS normalisation.
-    pub fn record_rms_norm(
-        &mut self,
-        batch: usize,
-        dim: usize,
-        eps: f32,
-    ) -> Result<(), MetalError> {
-        self.check_writable()?;
-        self.ops.push(CommandOp::RmsNorm { batch, dim, eps });
-        Ok(())
-    }
-
-    /// Record a RoPE embedding.
-    pub fn record_rope(
-        &mut self,
-        seq_len: usize,
-        head_dim: usize,
-        base_theta: f32,
-    ) -> Result<(), MetalError> {
-        self.check_writable()?;
-        self.ops
-            .push(CommandOp::RoPE {
-                seq_len,
-                head_dim,
-                base_theta,
-            });
-        Ok(())
-    }
-
-    /// Record a device-to-device copy.
-    pub fn record_copy(&mut self, num_bytes: usize) -> Result<(), MetalError> {
-        self.check_writable()?;
-        self.ops.push(CommandOp::Copy { num_bytes });
-        Ok(())
-    }
-
-    /// Number of recorded ops.
-    pub fn op_count(&self) -> usize {
-        self.ops.len()
-    }
-
-    /// Total estimated FLOPs across all recorded ops.
-    pub fn total_flops(&self) -> u64 {
-        self.ops.iter().map(|op| op.estimated_flops()).sum()
-    }
-
-    /// Seal the encoder, producing a frozen `CommandBuffer`.
-    ///
-    /// After sealing, no more ops can be recorded.
-    pub fn seal(mut self) -> Result<CommandBuffer, MetalError> {
-        if self.sealed {
-            return Err(MetalError::CommandSubmission(
-                "Encoder already sealed".to_string(),
-            ));
-        }
-        if self.ops.is_empty() {
-            return Err(MetalError::CommandSubmission(
-                "Cannot seal empty command buffer".to_string(),
-            ));
-        }
-        self.sealed = true;
-        Ok(CommandBuffer {
-            ops: std::mem::take(&mut self.ops),
-            label: self.label.clone(),
-        })
-    }
-
-    fn check_writable(&self) -> Result<(), MetalError> {
-        if self.sealed {
-            return Err(MetalError::CommandSubmission(
-                "Cannot record into sealed encoder".to_string(),
-            ));
-        }
-        Ok(())
-    }
-}
-
-impl fmt::Debug for CommandBufferEncoder {
-    fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
-        f.debug_struct("CommandBufferEncoder")
-            .field("label", &self.label)
-            .field("ops", &self.ops.len())
-            .field("sealed", &self.sealed)
-            .finish()
-    }
-}
-
-// ---------------------------------------------------------------------------
-// Sealed Command Buffer
-// ---------------------------------------------------------------------------
-
-/// A frozen, immutable list of compute operations ready for dispatch.
-///
-/// Created by `CommandBufferEncoder::seal()`. Cannot be modified after creation.
-pub struct CommandBuffer {
-    ops: Vec<CommandOp>,
-    label: String,
-}
-
-impl CommandBuffer {
-    /// The operations in submission order.
-    pub fn ops(&self) -> &[CommandOp] {
-        &self.ops
-    }
-
-    /// Debug label.
-    pub fn label(&self) -> &str {
-        &self.label
-    }
-
-    /// Number of ops.
-    pub fn len(&self) -> usize {
-        self.ops.len()
-    }
-
-    pub fn is_empty(&self) -> bool {
-        self.ops.is_empty()
-    }
-
-    /// Total estimated FLOPs.
-    pub fn total_flops(&self) -> u64 {
-        self.ops.iter().map(|op| op.estimated_flops()).sum()
-    }
-}
-
-impl fmt::Debug for CommandBuffer {
-    fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
-        f.debug_struct("CommandBuffer")
-            .field("label", &self.label)
-            .field("ops", &self.ops.len())
-            .field("total_flops", &self.total_flops())
-            .finish()
-    }
-}
-
-// ---------------------------------------------------------------------------
-// Compute Backend
-// ---------------------------------------------------------------------------
-
-/// Compute backend for dispatching operations.
-#[derive(Debug, Clone, Copy, PartialEq, Eq, Hash)]
-pub enum ComputeBackend {
-    /// Apple Metal GPU (highest priority on macOS).
-    Metal,
-    /// CPU fallback (always available).
-    Cpu,
-    /// Remote offload (future: network-attached accelerator).
-    Remote,
-}
-
-impl ComputeBackend {
-    /// Backend dispatch priority (lower = higher priority).
-    pub fn priority(&self) -> u8 {
-        match self {
-            ComputeBackend::Metal => 0,
-            ComputeBackend::Remote => 1,
-            ComputeBackend::Cpu => 2,
-        }
-    }
-
-    /// Human-readable name.
-    pub fn name(&self) -> &'static str {
-        match self {
-            ComputeBackend::Metal => "metal",
-            ComputeBackend::Cpu => "cpu",
-            ComputeBackend::Remote => "remote",
-        }
-    }
-}
-
-// ---------------------------------------------------------------------------
-// Execution Status + Handle
-// ---------------------------------------------------------------------------
-
-/// Status of a submitted command buffer.
-#[derive(Debug, Clone, Copy, PartialEq, Eq)]
-pub enum ExecutionStatus {
-    /// Queued but not yet started.
-    Pending,
-    /// Currently executing on a backend.
-    Running,
-    /// Completed successfully.
-    Completed,
-    /// Failed with an error.
-    Failed,
-}
-
-/// Handle for tracking async command buffer execution.
-///
-/// Returned by `TieredExecutor::submit()`.
-pub struct ExecutionHandle {
-    pub id: u64,
-    pub status: ExecutionStatus,
-    pub backend: ComputeBackend,
-    pub label: String,
-    pub ops_completed: usize,
-    pub ops_total: usize,
-}
-
-impl ExecutionHandle {
-    /// Check if execution is finished (completed or failed).
-    pub fn is_done(&self) -> bool {
-        matches!(
-            self.status,
-            ExecutionStatus::Completed | ExecutionStatus::Failed
-        )
-    }
-
-    /// Progress as a fraction ∈ [0, 1].
-    pub fn progress(&self) -> f32 {
-        if self.ops_total == 0 {
-            return 1.0;
-        }
-        self.ops_completed as f32 / self.ops_total as f32
-    }
-}
-
-impl fmt::Debug for ExecutionHandle {
-    fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
-        f.debug_struct("ExecutionHandle")
-            .field("id", &self.id)
-            .field("status", &self.status)
-            .field("backend", &self.backend)
-            .field("progress", &format!("{:.0}%", self.progress() * 100.0))
-            .finish()
-    }
-}
-
-// ---------------------------------------------------------------------------
-// Tiered Executor
-// ---------------------------------------------------------------------------
-
-/// Dispatches sealed `CommandBuffer`s to the best available backend.
-///
-/// Backend selection order: Metal → Remote → CPU.
-/// Each submission returns an `ExecutionHandle` for polling.
-///
-/// In production, this manages MTLCommandQueue submission and async
-/// completion handlers. The current implementation executes synchronously
-/// on CPU as a functional stub.
-pub struct TieredExecutor {
-    /// Available backends, sorted by priority (best first).
-    backends: Vec<ComputeBackend>,
-    /// Next handle ID.
-    next_id: u64,
-    /// Total ops executed (lifetime counter).
-    total_ops_dispatched: u64,
-    /// Total FLOPs dispatched (lifetime counter).
-    total_flops_dispatched: u64,
-}
-
-impl TieredExecutor {
-    /// Create with a set of available backends.
-    ///
-    /// Backends are sorted by priority automatically.
-    pub fn new(mut backends: Vec<ComputeBackend>) -> Self {
-        backends.sort_by_key(|b| b.priority());
-        backends.dedup();
-        Self {
-            backends,
-            next_id: 0,
-            total_ops_dispatched: 0,
-            total_flops_dispatched: 0,
-        }
-    }
-
-    /// Create with only CPU backend (always available).
-    pub fn cpu_only() -> Self {
-        Self::new(vec![ComputeBackend::Cpu])
-    }
-
-    /// Submit a sealed command buffer for execution.
-    ///
-    /// Selects the highest-priority available backend and executes.
-    /// Returns an `ExecutionHandle` reflecting final status.
-    pub fn submit(&mut self, cmd: &CommandBuffer) -> MetalResult<ExecutionHandle> {
-        let backend = self.backends.first().copied().ok_or_else(|| {
-            MetalError::CommandSubmission("No compute backends available".to_string())
-        })?;
-
-        let id = self.next_id;
-        self.next_id += 1;
-        let ops_total = cmd.len();
-
-        // Execute all ops (synchronous simulation)
-        let mut ops_completed = 0;
-        for op in cmd.ops() {
-            // In production: encode op into MTLCommandBuffer
-            // For now: just count
-            self.total_flops_dispatched += op.estimated_flops();
-            ops_completed += 1;
-        }
-        self.total_ops_dispatched += ops_total as u64;
-
-        Ok(ExecutionHandle {
-            id,
-            status: ExecutionStatus::Completed,
-            backend,
-            label: cmd.label().to_string(),
-            ops_completed,
-            ops_total,
-        })
-    }
-
-    /// Number of available backends.
-    pub fn backend_count(&self) -> usize {
-        self.backends.len()
-    }
-
-    /// The highest-priority backend.
-    pub fn primary_backend(&self) -> Option<ComputeBackend> {
-        self.backends.first().copied()
-    }
-
-    /// Total ops dispatched over the executor's lifetime.
-    pub fn total_ops_dispatched(&self) -> u64 {
-        self.total_ops_dispatched
-    }
-
-    /// Total FLOPs dispatched over the executor's lifetime.
-    pub fn total_flops_dispatched(&self) -> u64 {
-        self.total_flops_dispatched
-    }
-
-    /// Check if a specific backend is available.
-    pub fn has_backend(&self, backend: ComputeBackend) -> bool {
-        self.backends.contains(&backend)
-    }
-}
-
-impl fmt::Debug for TieredExecutor {
-    fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
-        f.debug_struct("TieredExecutor")
-            .field("backends", &self.backends)
-            .field("total_ops", &self.total_ops_dispatched)
-            .field("total_flops", &self.total_flops_dispatched)
-            .finish()
-    }
-}
-
-// ---------------------------------------------------------------------------
-// Tests
-// ---------------------------------------------------------------------------
-
-#[cfg(test)]
-mod tests {
-    use super::*;
-
-    #[test]
-    fn test_metal_dtype_sizes() {
-        assert_eq!(MetalDtype::Float32.size_bytes(), 4);
-        assert_eq!(MetalDtype::Float16.size_bytes(), 2);
-        assert_eq!(MetalDtype::BFloat16.size_bytes(), 2);
-        assert_eq!(MetalDtype::Int8.size_bytes(), 1);
-        assert_eq!(MetalDtype::Int32.size_bytes(), 4);
-        assert_eq!(MetalDtype::Uint8.size_bytes(), 1);
-        assert_eq!(MetalDtype::Uint32.size_bytes(), 4);
-    }
-
-    #[test]
-    fn test_metal_dtype_msl_names() {
-        assert_eq!(MetalDtype::Float32.msl_type(), "float");
-        assert_eq!(MetalDtype::Float16.msl_type(), "half");
-        assert_eq!(MetalDtype::Int8.msl_type(), "char");
-    }
-
-    #[test]
-    fn test_buffer_creation() {
-        let buf = MetalBuffer::new(1024, MetalDtype::Float32, 1).unwrap();
-        assert_eq!(buf.element_count, 1024);
-        assert_eq!(buf.size_bytes, 4096);
-        assert_eq!(buf.dtype, MetalDtype::Float32);
-    }
-
-    #[test]
-    fn test_buffer_zero_size_rejected() {
-        let result = MetalBuffer::new(0, MetalDtype::Float32, 1);
-        assert!(result.is_err());
-    }
-
-    #[test]
-    fn test_buffer_from_f32_roundtrip() {
-        let data = vec![1.0f32, 2.0, 3.0, 4.0];
-        let buf = MetalBuffer::from_f32(&data, 1).unwrap();
-        let read = buf.as_f32().unwrap();
-        assert_eq!(read, &data[..]);
-    }
-
-    #[test]
-    fn test_buffer_wrong_dtype_read() {
-        let buf = MetalBuffer::new(10, MetalDtype::Int8, 1).unwrap();
-        assert!(buf.as_f32().is_err());
-    }
-
-    #[test]
-    fn test_buffer_write_f32() {
-        let mut buf = MetalBuffer::new(4, MetalDtype::Float32, 1).unwrap();
-        buf.write_f32(&[5.0, 6.0, 7.0, 8.0]).unwrap();
-        let read = buf.as_f32().unwrap();
-        assert_eq!(read, &[5.0, 6.0, 7.0, 8.0]);
-    }
-
-    #[test]
-    fn test_buffer_write_overflow() {
-        let mut buf = MetalBuffer::new(2, MetalDtype::Float32, 1).unwrap();
-        let result = buf.write_f32(&[1.0, 2.0, 3.0]); // 3 elements into 2-element buffer
-        assert!(result.is_err());
-    }
-
-    #[test]
-    fn test_threadgroup_size() {
-        let tg = ThreadgroupSize::new_1d(256);
-        assert_eq!(tg.total_threads(), 256);
-
-        let tg2 = ThreadgroupSize::new_2d(16, 16);
-        assert_eq!(tg2.total_threads(), 256);
-    }
-
-    #[test]
-    fn test_kernel_creation_and_compile() {
-        let mut kernel = MetalKernel::new(
-            "test_kernel",
-            "kernel void test() {}",
-            ThreadgroupSize::new_1d(64),
-        );
-        assert!(!kernel.is_compiled);
-        kernel.compile().unwrap();
-        assert!(kernel.is_compiled);
-    }
-
-    #[test]
-    fn test_kernel_empty_source_fails() {
-        let mut kernel = MetalKernel::new("bad", "", ThreadgroupSize::new_1d(64));
-        assert!(kernel.compile().is_err());
-    }
-
-    #[test]
-    fn test_kernel_grid_computation() {
-        let kernel = MetalKernel::new("k", "src", ThreadgroupSize::new_1d(256));
-        let grid = kernel.grid_for_elements(1000);
-        assert!(grid.width >= 1000);
-        assert_eq!(grid.width % 256, 0);
-    }
-
-    #[test]
-    fn test_context_creation() {
-        let ctx = MetalContext::new().unwrap();
-        assert!(ctx.device_name.contains("simulated"));
-        assert_eq!(ctx.allocated_bytes(), 0);
-        assert_eq!(ctx.kernel_count(), 0);
-    }
-
-    #[test]
-    fn test_context_buffer_lifecycle() {
-        let mut ctx = MetalContext::new().unwrap();
-
-        let buf = ctx.alloc_buffer(1024, MetalDtype::Float32).unwrap();
-        assert_eq!(ctx.allocated_bytes(), 4096);
-
-        ctx.free_buffer(buf);
-        assert_eq!(ctx.allocated_bytes(), 0);
-    }
-
-    #[test]
-    fn test_context_alloc_f32() {
-        let mut ctx = MetalContext::new().unwrap();
-        let data = vec![1.0, 2.0, 3.0];
-        let buf = ctx.alloc_f32(&data).unwrap();
-        assert_eq!(buf.as_f32().unwrap(), &[1.0, 2.0, 3.0]);
-        assert_eq!(ctx.allocated_bytes(), 12);
-    }
-
-    #[test]
-    fn test_context_register_kernel() {
-        let mut ctx = MetalContext::new().unwrap();
-        ctx.register_kernel("add", "kernel void add() {}", ThreadgroupSize::new_1d(64))
-            .unwrap();
-        assert_eq!(ctx.kernel_count(), 1);
-        assert!(ctx.get_kernel("add").is_some());
-        assert!(ctx.get_kernel("missing").is_none());
-    }
-
-    #[test]
-    fn test_context_dispatch_unregistered_fails() {
-        let ctx = MetalContext::new().unwrap();
-        let result = ctx.dispatch("unknown_k", GridSize::new_1d(64), &[]);
-        assert!(result.is_err());
-    }
-
-    #[test]
-    fn test_context_dispatch_zero_grid_fails() {
-        let mut ctx = MetalContext::new().unwrap();
-        ctx.register_kernel("k", "src", ThreadgroupSize::new_1d(64)).unwrap();
-        let result = ctx.dispatch("k", GridSize { width: 0, height: 1, depth: 1 }, &[]);
-        assert!(result.is_err());
-    }
-
-    #[test]
-    fn test_elementwise_add() {
-        let mut ctx = MetalContext::new().unwrap();
-        let a = vec![1.0, 2.0, 3.0, 4.0];
-        let b = vec![10.0, 20.0, 30.0, 40.0];
-        let result = ctx.elementwise_add(&a, &b).unwrap();
-        assert_eq!(result, vec![11.0, 22.0, 33.0, 44.0]);
-        // All temp buffers freed
-        assert_eq!(ctx.allocated_bytes(), 0);
-    }
-
-    #[test]
-    fn test_elementwise_add_mismatched_lengths() {
-        let mut ctx = MetalContext::new().unwrap();
-        let result = ctx.elementwise_add(&[1.0, 2.0], &[1.0]);
-        assert!(result.is_err());
-    }
-
-    #[test]
-    fn test_kernel_library_sources_not_empty() {
-        assert!(!MetalKernelLibrary::softmax_source().is_empty());
-        assert!(!MetalKernelLibrary::rmsnorm_source().is_empty());
-        assert!(!MetalKernelLibrary::silu_source().is_empty());
-        assert!(!MetalKernelLibrary::rope_source().is_empty());
-    }
-
-    #[test]
-    fn test_kernel_library_sources_compile() {
-        let mut ctx = MetalContext::new().unwrap();
-
-        ctx.register_kernel("softmax", MetalKernelLibrary::softmax_source(), ThreadgroupSize::new_1d(256))
-            .unwrap();
-        ctx.register_kernel("rmsnorm", MetalKernelLibrary::rmsnorm_source(), ThreadgroupSize::new_1d(256))
-            .unwrap();
-        ctx.register_kernel("silu", MetalKernelLibrary::silu_source(), ThreadgroupSize::new_1d(256))
-            .unwrap();
-        ctx.register_kernel("rope", MetalKernelLibrary::rope_source(), ThreadgroupSize::new_1d(256))
-            .unwrap();
-
-        assert_eq!(ctx.kernel_count(), 4);
-    }
-
-    #[test]
-    fn test_metal_error_display() {
-        let err = MetalError::DeviceNotFound;
-        assert_eq!(err.to_string(), "No Metal-capable GPU found");
-
-        let err = MetalError::BufferTooSmall {
-            required: 100,
-            actual: 50,
-        };
-        assert!(err.to_string().contains("100"));
-        assert!(err.to_string().contains("50"));
-    }
-
-    #[test]
-    fn test_multiple_buffers_accounting() {
-        let mut ctx = MetalContext::new().unwrap();
-
-        let b1 = ctx.alloc_buffer(100, MetalDtype::Float32).unwrap(); // 400 bytes
-        let b2 = ctx.alloc_buffer(200, MetalDtype::Float16).unwrap(); // 400 bytes
-        assert_eq!(ctx.allocated_bytes(), 800);
-
-        ctx.free_buffer(b1);
-        assert_eq!(ctx.allocated_bytes(), 400);
-
-        ctx.free_buffer(b2);
-        assert_eq!(ctx.allocated_bytes(), 0);
-    }
-
-    // -----------------------------------------------------------------------
-    // Execution subsystem tests
-    // -----------------------------------------------------------------------
-
-    #[test]
-    fn test_command_op_names() {
-        let matmul = CommandOp::MatMul { rows_a: 1, inner: 1, cols_b: 1, dtype: MetalDtype::Float32 };
-        assert_eq!(matmul.name(), "matmul");
-        assert_eq!(CommandOp::Softmax { rows: 1, row_size: 1 }.name(), "softmax");
-        assert_eq!(CommandOp::LayerNorm { batch: 1, dim: 1, eps: 1e-5 }.name(), "layer_norm");
-        assert_eq!(CommandOp::RmsNorm { batch: 1, dim: 1, eps: 1e-5 }.name(), "rms_norm");
-        assert_eq!(CommandOp::RoPE { seq_len: 1, head_dim: 1, base_theta: 10000.0 }.name(), "rope");
-        assert_eq!(CommandOp::Copy { num_bytes: 1 }.name(), "copy");
-    }
-
-    #[test]
-    fn test_matmul_flops() {
-        // 2 * M * K * N
-        let op = CommandOp::MatMul { rows_a: 32, inner: 4096, cols_b: 4096, dtype: MetalDtype::Float16 };
-        assert_eq!(op.estimated_flops(), 2 * 32 * 4096 * 4096);
-    }
-
-    #[test]
-    fn test_encoder_record_all_ops() {
-        let mut enc = CommandBufferEncoder::new("layer_0");
-        enc.record_matmul(32, 4096, 4096, MetalDtype::Float32).unwrap();
-        enc.record_softmax(32, 128).unwrap();
-        enc.record_layer_norm(32, 4096, 1e-5).unwrap();
-        enc.record_rms_norm(32, 4096, 1e-5).unwrap();
-        enc.record_rope(128, 64, 10000.0).unwrap();
-        enc.record_copy(1024).unwrap();
-        assert_eq!(enc.op_count(), 6);
-        assert!(enc.total_flops() > 0);
-    }
-
-    #[test]
-    fn test_encoder_seal_produces_buffer() {
-        let mut enc = CommandBufferEncoder::new("test");
-        enc.record_softmax(8, 256).unwrap();
-        let buf = enc.seal().unwrap();
-        assert_eq!(buf.len(), 1);
-        assert_eq!(buf.label(), "test");
-        assert!(!buf.is_empty());
-    }
-
-    #[test]
-    fn test_seal_empty_encoder_fails() {
-        let enc = CommandBufferEncoder::new("empty");
-        let result = enc.seal();
-        assert!(result.is_err());
-    }
-
-    #[test]
-    fn test_command_buffer_flops() {
-        let mut enc = CommandBufferEncoder::new("multi");
-        enc.record_matmul(1, 100, 100, MetalDtype::Float32).unwrap();
-        enc.record_copy(500).unwrap();
-        let buf = enc.seal().unwrap();
-        // matmul: 2*1*100*100 = 20000, copy: 500
-        assert_eq!(buf.total_flops(), 20500);
-    }
-
-    #[test]
-    fn test_command_buffer_ops_order() {
-        let mut enc = CommandBufferEncoder::new("order");
-        enc.record_matmul(1, 1, 1, MetalDtype::Float32).unwrap();
-        enc.record_softmax(1, 1).unwrap();
-        enc.record_rms_norm(1, 1, 1e-5).unwrap();
-        let buf = enc.seal().unwrap();
-        assert_eq!(buf.ops()[0].name(), "matmul");
-        assert_eq!(buf.ops()[1].name(), "softmax");
-        assert_eq!(buf.ops()[2].name(), "rms_norm");
-    }
-
-    #[test]
-    fn test_backend_priority_order() {
-        assert!(ComputeBackend::Metal.priority() < ComputeBackend::Remote.priority());
-        assert!(ComputeBackend::Remote.priority() < ComputeBackend::Cpu.priority());
-    }
-
-    #[test]
-    fn test_backend_names() {
-        assert_eq!(ComputeBackend::Metal.name(), "metal");
-        assert_eq!(ComputeBackend::Cpu.name(), "cpu");
-        assert_eq!(ComputeBackend::Remote.name(), "remote");
-    }
-
-    #[test]
-    fn test_executor_cpu_only() {
-        let exec = TieredExecutor::cpu_only();
-        assert_eq!(exec.backend_count(), 1);
-        assert_eq!(exec.primary_backend(), Some(ComputeBackend::Cpu));
-        assert!(exec.has_backend(ComputeBackend::Cpu));
-        assert!(!exec.has_backend(ComputeBackend::Metal));
-    }
-
-    #[test]
-    fn test_executor_auto_sorts_backends() {
-        let exec = TieredExecutor::new(vec![
-            ComputeBackend::Cpu,
-            ComputeBackend::Metal,
-            ComputeBackend::Remote,
-        ]);
-        // Metal should be first (priority 0)
-        assert_eq!(exec.primary_backend(), Some(ComputeBackend::Metal));
-        assert_eq!(exec.backend_count(), 3);
-    }
-
-    #[test]
-    fn test_executor_deduplicates() {
-        let exec = TieredExecutor::new(vec![
-            ComputeBackend::Cpu,
-            ComputeBackend::Cpu,
-            ComputeBackend::Cpu,
-        ]);
-        assert_eq!(exec.backend_count(), 1);
-    }
-
-    #[test]
-    fn test_executor_submit_completes() {
-        let mut exec = TieredExecutor::cpu_only();
-        let mut enc = CommandBufferEncoder::new("submit_test");
-        enc.record_softmax(4, 64).unwrap();
-        enc.record_rms_norm(4, 128, 1e-6).unwrap();
-        let buf = enc.seal().unwrap();
-        let handle = exec.submit(&buf).unwrap();
-        assert_eq!(handle.status, ExecutionStatus::Completed);
-        assert_eq!(handle.ops_completed, 2);
-        assert_eq!(handle.ops_total, 2);
-        assert!(handle.is_done());
-    }
-
-    #[test]
-    fn test_executor_handle_progress() {
-        let mut exec = TieredExecutor::cpu_only();
-        let mut enc = CommandBufferEncoder::new("prog");
-        enc.record_copy(100).unwrap();
-        let buf = enc.seal().unwrap();
-        let handle = exec.submit(&buf).unwrap();
-        assert!((handle.progress() - 1.0).abs() < f32::EPSILON);
-    }
-
-    #[test]
-    fn test_executor_lifetime_counters() {
-        let mut exec = TieredExecutor::cpu_only();
-
-        let mut enc1 = CommandBufferEncoder::new("b1");
-        enc1.record_matmul(2, 4, 8, MetalDtype::Float32).unwrap();
-        let buf1 = enc1.seal().unwrap();
-        exec.submit(&buf1).unwrap();
-
-        let mut enc2 = CommandBufferEncoder::new("b2");
-        enc2.record_copy(1000).unwrap();
-        enc2.record_softmax(2, 32).unwrap();
-        let buf2 = enc2.seal().unwrap();
-        exec.submit(&buf2).unwrap();
-
-        assert_eq!(exec.total_ops_dispatched(), 3); // 1 + 2
-        assert!(exec.total_flops_dispatched() > 0);
-    }
-
-    #[test]
-    fn test_executor_sequential_ids() {
-        let mut exec = TieredExecutor::cpu_only();
-        let mut ids = Vec::new();
-        for i in 0..5 {
-            let mut enc = CommandBufferEncoder::new(&format!("seq_{}", i));
-            enc.record_copy(1).unwrap();
-            let buf = enc.seal().unwrap();
-            let h = exec.submit(&buf).unwrap();
-            ids.push(h.id);
-        }
-        assert_eq!(ids, vec![0, 1, 2, 3, 4]);
-    }
-
-    #[test]
-    fn test_executor_selects_metal_over_cpu() {
-        let mut exec = TieredExecutor::new(vec![ComputeBackend::Cpu, ComputeBackend::Metal]);
-        let mut enc = CommandBufferEncoder::new("prio");
-        enc.record_copy(1).unwrap();
-        let buf = enc.seal().unwrap();
-        let handle = exec.submit(&buf).unwrap();
-        assert_eq!(handle.backend, ComputeBackend::Metal);
-    }
-
-    #[test]
-    fn test_execution_handle_done_states() {
-        let completed = ExecutionHandle {
-            id: 0,
-            status: ExecutionStatus::Completed,
-            backend: ComputeBackend::Cpu,
-            label: "a".to_string(),
-            ops_completed: 1,
-            ops_total: 1,
-        };
-        assert!(completed.is_done());
-
-        let failed = ExecutionHandle {
-            id: 1,
-            status: ExecutionStatus::Failed,
-            backend: ComputeBackend::Cpu,
-            label: "b".to_string(),
-            ops_completed: 0,
-            ops_total: 1,
-        };
-        assert!(failed.is_done());
-
-        let pending = ExecutionHandle {
-            id: 2,
-            status: ExecutionStatus::Pending,
-            backend: ComputeBackend::Cpu,
-            label: "c".to_string(),
-            ops_completed: 0,
-            ops_total: 1,
-        };
-        assert!(!pending.is_done());
-    }
-
-    #[test]
-    fn test_execution_handle_zero_ops_progress() {
-        let h = ExecutionHandle {
-            id: 0,
-            status: ExecutionStatus::Completed,
-            backend: ComputeBackend::Cpu,
-            label: "empty".to_string(),
-            ops_completed: 0,
-            ops_total: 0,
-        };
-        assert!((h.progress() - 1.0).abs() < f32::EPSILON);
-    }
-
-    #[test]
-    fn test_encoder_debug_format() {
-        let enc = CommandBufferEncoder::new("dbg");
-        let debug = format!("{:?}", enc);
-        assert!(debug.contains("dbg"));
-        assert!(debug.contains("CommandBufferEncoder"));
-    }
-}
+//! This shim keeps existing `use crate::ucal::*` imports compiling.
+//! Migrate to the direct module paths when convenient. See ADR-0005.
+
+pub use crate::shared_buffer::{ComputeBackend, SharedBuffer};
+pub use crate::metal_compute::{
+    CommandBuffer, CommandBufferEncoder, CommandOp, ExecutionHandle, ExecutionStatus,
+    GridSize, MetalBuffer, MetalContext, MetalDtype, MetalError, MetalKernel,
+    MetalKernelLibrary, MetalResult, ThreadgroupSize, TieredExecutor,
+};
